requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
mysql-connector-python==9.0.0
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4
